Switch position statement to use Uint32
Position is restricted to 32bit unsigned, there is just no point
to say it is a Long.

JIRA: YANGTOOLS-1065

+++ yang/yang-model-api/src/main/java/org/opendaylight/yangtools/yang/model/api/stmt/PositionEffectiveStatement.java
@@ -8,9 +8,10 @@
 package org.opendaylight.yangtools.yang.model.api.stmt;
 
 import com.google.common.annotations.Beta;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.model.api.meta.EffectiveStatement;
 
 @Beta
-public interface PositionEffectiveStatement extends EffectiveStatement<Long, PositionStatement> {
+public interface PositionEffectiveStatement extends EffectiveStatement<Uint32, PositionStatement> {
 
 }
